fix positiveBtn 추가 수정
